requirements.txt
streamlit==1.47.0
langchain==0.3.26
langchain-community==0.3.27
langchain-google-genai==2.1.8
python-dotenv==1.1.1
psycopg2-binary==2.9.10
langchain-openai==0.3.28
langchain-huggingface==0.3.0
langchain-experimental==0.3.4
